+++ dapp-installer.linux.config.json
@@ -1,0 +1,5 @@
+{
+    "Role" : "agent",
+    "Path" : "/home/parallels/agent",
+    "Source":"/tmp/2874d0528f35b8135bacf4827f86887e/common.tar.xz"
+ }
